requirements.txt
Flask==1.1.2
jinja2==2.11.3
werkzeug==0.16.1
markupsafe==1.1.1
itsdangerous==1.1.0
google-generativeai
python-dotenv
pandas
pymongo
pyspark
numpy
scikit-learn
faiss-cpu
psutil  
py4j  
